vendor: Bump controller-tools to latest
This bump brings controller-tools up-to-date with the upstream master
branch as of Sept 6, 2020, and support for adding a top-level
"description" field to the validation schema. This allows CNPs and CCNPs
to have a human-readable description, useful for explaining what those
policies represent.

When applying a CNP or a CCNP with a description field (see example
below), having this "description" field present in the validation schema
is required. It was previously not required, as v1beta1 CRDs permitted
"unknown" fields. Moving to v1 CRDs disallows "unknown" fields
(`preserveUnknownFields`), hence why the support was added and updated
via this commit.

Example CNP / CCNP:

```
apiVersion: "cilium.io/v2"
kind: CiliumNetworkPolicy
description: "L7 policy for getting started using Kubernetes guide"
metadata:
  name: "l7-policy"
spec:
  endpointSelector:
    matchLabels:
      id: app1
  ingress:
  - fromEndpoints:
    - matchLabels:
        id: app2
    toPorts:
    - ports:
      - port: "80"
        protocol: TCP
      rules:
        http:
        - method: "GET"
          path: "/public"
```

Signed-off-by: Chris Tarazi <chris@isovalent.com>